requirements.txt
gymnasium
stable-baselines3
numpy
pandas
matplotlib
yfinance
Optuna